edit group function

+++ app/src/main/java/com/csis4495_cmk/webuy/models/Group.java
@@ -136,7 +136,7 @@ public class Group {
     }
 
 
-    public double getTax() {
+    public int getTax() {
         return tax;
     }
